updates to acr pipeline

+++ scripts/dpo/iter_dpo_main.sh
@@ -22,13 +22,15 @@ openrlhf.cli.train_dpo \
     --train_batch_size 128 \
     --gradient_checkpointing \
     --seed 123456 \
-    --zero_stage 2 \
+    --zero_stage 3 \
+    --adam_offload \
     --bf16 \
     --learning_rate 1e-4 \
     --beta 0.15 \
     --adam_betas 0.9 0.98 \
     --max_epochs 1 \
     --pretrain /workspace/models/gemma-2-9b-GENERATOR \
+    --ref_pretrain /workspace/models/gemma-2-9b-prev \
     --dataset /workspace/CharacterTraining/data/current_dpo.jsonl \
     --chosen_key revision \
     --rejected_key initial \
@@ -40,15 +42,20 @@ openrlhf.cli.train_dpo \
 EOF
     deepspeed \
     --module $training_commands
+    if [ $? -ne 0 ]; then
+        echo "error: deepspeed command failed in iteration $i"
+        exit 1
+    fi
     # remove wandb folder
     rm -rf /workspace/wandb
     # upload the new model
     cd /workspace/CharacterTraining/tools
     python upload_model.py \
         --model gemma-2-9b-next \
-        --name gemma-2-9b-next-0703-iter-$i
-    # build the snapshot for the next generation step (first remove the old snapshot)
-    rm -rf /workspace/models/gemma-2-9b-GENERATOR
+        --name gemma-2-9b-next-0803-iter-$i
+    # build the snapshot for the next generation step 
+    rm -rf /workspace/models/gemma-2-9b-prev
+    mv /workspace/models/gemma-2-9b-GENERATOR /workspace/models/gemma-2-9b-prev
     mv /workspace/models/gemma-2-9b-next /workspace/models/gemma-2-9b-GENERATOR
     echo "finished DPO iteration $i of 5"
 done
